+++ zrml/neo-swaps/src/tests/deploy_pool.rs
@@ -262,7 +262,7 @@ fn deploy_pool_fails_on_asset_count_above_max() {
         let market_id = create_market(
             ALICE,
             BASE_ASSET,
-            MarketType::Categorical(category_count),
+            MarketType::Categorical(category_count as u16),
             ScoringRule::Lmsr,
         );
         let liquidity = _10;
